Final project done
Card game project made as one of the XAMX university course challenges.

Player can interact with:
- own cards
- own portrait
- enemy portrait
- burn pile

Score increases for each card played and enemy defeated.
There are 2 enemy types (fire/ice)
Spells are Fire, Ice and Fire/Ice combined so they deal less or equal damage as listed based on enemy type.
There is a destruction spell which kills the enemy or yourself if you cast the spell while the enemy has a mirror effect. But spells can also bounce back and forth if both the player and the enemy have a mirror effect in play.